# Improve N3Store lookup efficiency by maintaining an inverse index