--- a/lib/N3Store.js
+++ b/lib/N3Store.js
@@ -89,7 +89,9 @@ N3Store.prototype = {
   // (for instance: _subject_, _predicate_, and _object_).
   // Finally, `graph` will be the graph of the created triples.
   _findInIndex: function (index0, key0, key1, key2, name0, name1, name2, graph) {
-    var results = [], entityKeys = this._entitiesInverse, tmp, index1, index2;
+    var vars = (key0 ? 1 : 0) + (key1 ? 1 : 0) + (key2 ? 1 : 0);
+    var results = [], entityKeys = (vars > 1 && !graph) ? Object.keys(this._entities) : this._entitiesInverse,
+        tmp, index1, index2;
 
     // If a key is specified, use only that part of index 0.
     if (key0) (tmp = index0, index0 = {})[key0] = tmp[key0];