--- a/perf/N3Store-perf.js
+++ b/perf/N3Store-perf.js
@@ -24,7 +24,28 @@ console.timeEnd(TEST);
 
 console.log('* Memory usage for triples: ' + Math.round(process.memoryUsage().rss / 1024 / 1024) + 'MB');
 
-TEST = '- Finding all ' + dimCubed + ' triples to the default graph ' + dimSquared * 3 + ' times';
+TEST = '- Finding all ' + dimCubed + ' triples to the default graph ' + dimSquared * 1 + ' times (0 variables)';
+console.time(TEST);
+for (i = 0; i < dim; i++)
+  for (j = 0; j < dim; j++)
+    for (k = 0; k < dim; k++)
+      assert.equal(store.find(prefix + i, prefix + j, prefix + k).length, 1);
+console.timeEnd(TEST);
+
+TEST = '- Finding all ' + dimCubed + ' triples to the default graph ' + dimSquared * 2 + ' times (1 variable)';
+console.time(TEST);
+for (i = 0; i < dim; i++)
+  for (j = 0; j < dim; j++)
+    assert.equal(store.find(prefix + i, prefix + j, null).length, dim);
+for (i = 0; i < dim; i++)
+  for (j = 0; j < dim; j++)
+    assert.equal(store.find(prefix + i, null, prefix + j).length, dim);
+for (i = 0; i < dim; i++)
+  for (j = 0; j < dim; j++)
+    assert.equal(store.find(null, prefix + i, prefix + j).length, dim);
+console.timeEnd(TEST);
+
+TEST = '- Finding all ' + dimCubed + ' triples to the default graph ' + dimSquared * 3 + ' times (2 variables)';
 console.time(TEST);
 for (i = 0; i < dim; i++)
   assert.equal(store.find(prefix + i, null, null).length, dimSquared);