This is an improvement for the lookup-speed of the `N3Store`, as discussed in https://github.com/rubensworks/N3.js/commit/87cae2914751c93fdeb8c005b6c445d002b6f446.
Apparently, `Object.keys` became more expensive since one of the newer Node versions,
so this PR avoids calling this function every time on `this._entities` when doing lookups.
Instead, a inverse-index is maintained to avoid this operation.

I ran the performance tests to compare this new approach with the old one,
and it resulted in some interesting results:

```
2 variables for lookups (a.k.a original perf tests):
NEW:
    N3Store performance test
    - Adding 16777216 triples in the default graph: 18177.111ms
    * Memory usage for triples: 831MB
    - Finding all 16777216 triples to the default graph 196608 times: 6421.472ms

    - Adding 16777216 quads: 24892.776ms
    * Memory usage for quads: 676MB
    - Finding all 16777216 quads 1048576 times: 13141.350ms

OLD:
    N3Store performance test
    - Adding 16777216 triples in the default graph: 19465.032ms
    * Memory usage for triples: 808MB
    - Finding all 16777216 triples to the default graph 196608 times: 3206.058ms

    - Adding 16777216 quads: 24860.754ms
    * Memory usage for quads: 717MB
    - Finding all 16777216 quads 1048576 times: 19608.113ms
------------------------------------------------------------------------------------
1 variable for lookup:
NEW:
    - Finding all 16777216 triples to the default graph 65536 times: 1207.522ms
OLD:
    - Finding all 16777216 triples to the default graph 65536 times: 2133.597ms
------------------------------------------------------------------------------------
0 variables for lookup:
NEW:
    - Finding all 16777216 triples to the default graph 65536 times: 92602.621ms
OLD:
    - Finding all 16777216 triples to the default graph 65536 times: 283470.831ms
```

It appears that this new approach (precalculation of inverse index) is two times *slower* than the original approach for lookups with two variables (? ? o, ? p ?, s ? ?).
When increasing the selectivity of the lookups, this new approach becomes *faster* than the old one.
The new version is twice as fast as the old version for lookups with only one variables.
And it is three times faster than the old version for lookups without variables.

Like you have said before, the efficiency strongly depends on the access patterns.
So I'm not sure if you should accept this PR, since it makes lookups slower with two variables, but faster for the rest.
Maybe we should make two versions of the `N3Store` that are optimized for different cases.
Or maybe some more elaborate solution is possible.