@rubensworks 2016-05-23
As discussed last week, changing `_findInIndex` to always use the fastest variant might be the best solution here.
`Object.keys(this._entities)` for 2 or more variables and `this._entitiesInverse` for all other cases.

To be investigated...