@drewp 2016-06-05
This patch helps my case a lot.

I have ~800 entities in my store so far:
```
      t1 = Date.now()
      for i in [0..1000]
        @graph.floatValue(@uri, U(':originTime')) // calls findByIRI(s, p), where s and p are non-null
      log('took', Date.now() - t1)

timeline.js:390 took 14654 (14ms per call!)
timeline.js:390 took 14314
timeline.js:390 took 14183
timeline.js:390 took 14003
timeline.js:390 took 14284
```

With the patch:
```
timeline.js:390 took 35 (0.035ms per call)
timeline.js:390 took 34
timeline.js:390 took 35
timeline.js:390 took 42
timeline.js:390 took 36
```