@rubensworks 2016-06-05
I'll have some free time tomorrow, so I can a look at this then. It should be possible to make sure that lookups are optimal for all access patterns.