@rubensworks 2016-06-06
This new version makes sure that the inverse index is only called when there are less than 2 variables,
in all other cases the original `Object.keys(this._entities)` is called.

The results below show that the new code (this PR) is faster than the original code in all cases.
(Tested with `dim = 64`)

```
OLD:
    N3Store performance test
    - Adding 262144 triples in the default graph: 323.389ms
    * Memory usage for triples: 46MB
    - Finding all 262144 triples to the default graph 4096 times (0 variables): 1755.370ms
    - Finding all 262144 triples to the default graph 8192 times (1 variable): 322.146ms
    - Finding all 262144 triples to the default graph 12288 times (2 variables): 242.523ms

    - Adding 65536 quads: 105.001ms
    * Memory usage for quads: 67MB
    - Finding all 65536 quads 16384 times: 99.361ms
NEW:
    N3Store performance test
    - Adding 262144 triples in the default graph: 614.931ms
    * Memory usage for triples: 57MB
    - Finding all 262144 triples to the default graph 4096 times (0 variables): 1538.850ms
    - Finding all 262144 triples to the default graph 8192 times (1 variable): 100.025ms
    - Finding all 262144 triples to the default graph 12288 times (2 variables): 59.296ms

    - Adding 65536 quads: 134.271ms
    * Memory usage for quads: 76MB
    - Finding all 65536 quads 16384 times: 28.999ms
```

The inverse index does however cause an increase in memory usage and triple insertion execution times.
If this would be a problem, a config option could be added to completely disable the inverse index and fall back to the original version,
this would be good for use cases where most of the lookups are done with two or more variables.