@drewp 2016-06-08
Something may have been contaminating my test before, making the 'old' version run extra long, but the patch is still a win for that one line:

```
      t1 = Date.now()
      for l in [0..10000]
        originTime = @graph.floatValue(@uri, U(':originTime'))
      log('took', Date.now() - t1)  
```

(this is in https://github.com/drewp/light9/blob/master/light9/web/timeline.coffee#L233)

N3.js
timeline.js:361 took 4700
timeline.js:361 took 4697
timeline.js:361 took 4713
timeline.js:361 took 4671
timeline.js:361 took 4729
timeline.js:361 took 4823
...

with git pull origin pull/61/head
```
timeline.js:361 took 144
timeline.js:361 took 203
timeline.js:361 took 205
timeline.js:361 took 202
timeline.js:361 took 167
...
```

My project happens to have no less-selective queries yet, so I may be unaffected by the new code. I also have not run with data larger than a few hundred quads yet. There's going to be a lot of point-on-curve statements with very repetitive subjects and predicates.