@RubenVerborgh 2016-06-23
I don't see the insertion performance hit as a problem, given that lookups will typically be much more common than insertions.