@RubenVerborgh 2016-06-24
Does that mean we simply don't need the `Object.keys` solution anymore?