@RubenVerborgh 2016-06-24
(Reverted merge for now, but can't reopen pull request unfortunately.)