@rubensworks 2016-06-24
Ah yes, the `!graph` part was added last-minute, I didn't realize at that time that the graph could never be falsy there. That is why the test didn't fail for the missing `this._entities['><'] = 0;`.

It's still strange then that the performance test showed a speedup in all cases...
I'll look into it and open a new PR.