@RubenVerborgh 2016-06-24
> I didn't realize at that time that the graph could never be falsy there
The graph is always a string/IRI (even the default graph has one).

BTW You can check whether all code paths are reached through `npm run coverage`.